backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
google_search_results==2.4.2
greenlet==3.3.1
h11==0.16.0
idna==3.11
playwright==1.58.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
python-multipart==0.0.22
RapidFuzz==3.14.3
requests==2.32.5
starlette==0.50.0
thefuzz==0.22.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
python-dotenv